requirements.txt
matplotlib==3.8.2
numpy==1.26.2
pandas==2.1.3
psycopg2==2.9.9
python-dotenv==1.0.0
